refactor(workstation): extract commit-split actions into useCommitSplitActions (0.72 app.ts decomposition)
